BUG: Ensure All Events are Processed through the VR CustomProcessEvents Callback
Following the previous commit titled "BUG: Ensure passing of vtkEventData to
displayable managers," this update guarantees the effective invocation of the
newly introduced custom delegation function, designed to handle a calldata
argument. This ensures that all processing is funneled through the
`vtkVirtualRealityViewInteractorObserver::CustomProcessEvents` callback. The
callback is responsible for handling 3D event data and passing it down appropriately.